+++ tests/fl_interface.py
@@ -25,8 +25,6 @@ class FunctionCategory(str, Enum):
 class InputIdentifiers(str, Enum):
     local = "local"
     shared = "shared"
-    model = "model"
-    models = "models"
     predictions = "predictions"
     performance = "performance"
     opener = "opener"
@@ -36,7 +34,6 @@ class InputIdentifiers(str, Enum):
 class OutputIdentifiers(str, Enum):
     local = "local"
     shared = "shared"
-    model = "model"
     predictions = "predictions"
     performance = "performance"
 
@@ -45,7 +42,7 @@ class FLFunctionInputs(list, Enum):
     """Substra function inputs by function category based on the InputIdentifiers"""
 
     FUNCTION_AGGREGATE = [
-        FunctionInputSpec(identifier=InputIdentifiers.models, kind=AssetKind.model.value, optional=False, multiple=True)
+        FunctionInputSpec(identifier=InputIdentifiers.shared, kind=AssetKind.model.value, optional=False, multiple=True)
     ]
     FUNCTION_SIMPLE = [
         FunctionInputSpec(
@@ -57,7 +54,7 @@ class FLFunctionInputs(list, Enum):
         FunctionInputSpec(
             identifier=InputIdentifiers.opener, kind=AssetKind.data_manager.value, optional=False, multiple=False
         ),
-        FunctionInputSpec(identifier=InputIdentifiers.models, kind=AssetKind.model.value, optional=True, multiple=True),
+        FunctionInputSpec(identifier=InputIdentifiers.shared, kind=AssetKind.model.value, optional=True, multiple=True),
     ]
     FUNCTION_COMPOSITE = [
         FunctionInputSpec(
@@ -85,7 +82,7 @@ class FLFunctionInputs(list, Enum):
             identifier=InputIdentifiers.opener, kind=AssetKind.data_manager.value, optional=False, multiple=False
         ),
         FunctionInputSpec(
-            identifier=InputIdentifiers.model, kind=AssetKind.model.value, optional=False, multiple=False
+            identifier=InputIdentifiers.shared, kind=AssetKind.model.value, optional=False, multiple=False
         ),
     ]
     FUNCTION_PREDICT_COMPOSITE = [
@@ -125,10 +122,10 @@ class FLFunctionOutputs(list, Enum):
     """Substra function outputs by function category based on the OutputIdentifiers"""
 
     FUNCTION_AGGREGATE = [
-        FunctionOutputSpec(identifier=OutputIdentifiers.model, kind=AssetKind.model.value, multiple=False)
+        FunctionOutputSpec(identifier=OutputIdentifiers.shared, kind=AssetKind.model.value, multiple=False)
     ]
     FUNCTION_SIMPLE = [
-        FunctionOutputSpec(identifier=OutputIdentifiers.model, kind=AssetKind.model.value, multiple=False)
+        FunctionOutputSpec(identifier=OutputIdentifiers.shared, kind=AssetKind.model.value, multiple=False)
     ]
     FUNCTION_COMPOSITE = [
         FunctionOutputSpec(identifier=OutputIdentifiers.local, kind=AssetKind.model.value, multiple=False),
@@ -168,9 +165,9 @@ class FLTaskInputGenerator:
     def trains_to_train(model_keys):
         return [
             InputRef(
-                identifier=InputIdentifiers.models,
+                identifier=InputIdentifiers.shared,
                 parent_task_key=model_key,
-                parent_task_output_identifier=OutputIdentifiers.model,
+                parent_task_output_identifier=OutputIdentifiers.shared,
             )
             for model_key in model_keys
         ]
@@ -179,9 +176,9 @@ class FLTaskInputGenerator:
     def trains_to_aggregate(model_keys):
         return [
             InputRef(
-                identifier=InputIdentifiers.models,
+                identifier=InputIdentifiers.shared,
                 parent_task_key=model_key,
-                parent_task_output_identifier=OutputIdentifiers.model,
+                parent_task_output_identifier=OutputIdentifiers.shared,
             )
             for model_key in model_keys
         ]
@@ -190,9 +187,9 @@ class FLTaskInputGenerator:
     def train_to_predict(model_key):
         return [
             InputRef(
-                identifier=InputIdentifiers.model,
+                identifier=InputIdentifiers.shared,
                 parent_task_key=model_key,
-                parent_task_output_identifier=OutputIdentifiers.model,
+                parent_task_output_identifier=OutputIdentifiers.shared,
             )
         ]
 
@@ -252,7 +249,7 @@ class FLTaskInputGenerator:
             InputRef(
                 identifier=InputIdentifiers.shared,
                 parent_task_key=model_key,
-                parent_task_output_identifier=OutputIdentifiers.model,
+                parent_task_output_identifier=OutputIdentifiers.shared,
             )
         ]
 
@@ -260,7 +257,7 @@ class FLTaskInputGenerator:
     def composites_to_aggregate(model_keys):
         return [
             InputRef(
-                identifier=InputIdentifiers.models,
+                identifier=InputIdentifiers.shared,
                 parent_task_key=model_key,
                 parent_task_output_identifier=OutputIdentifiers.shared,
             )
@@ -271,9 +268,9 @@ class FLTaskInputGenerator:
     def aggregate_to_predict(model_key):
         return [
             InputRef(
-                identifier=InputIdentifiers.models,
+                identifier=InputIdentifiers.shared,
                 parent_task_key=model_key,
-                parent_task_output_identifier=OutputIdentifiers.model,
+                parent_task_output_identifier=OutputIdentifiers.shared,
             )
         ]
 
@@ -281,7 +278,7 @@ class FLTaskInputGenerator:
     def local_to_aggregate(model_key):
         return [
             InputRef(
-                identifier=InputIdentifiers.models,
+                identifier=InputIdentifiers.shared,
                 parent_task_key=model_key,
                 parent_task_output_identifier=OutputIdentifiers.local,
             )
@@ -291,7 +288,7 @@ class FLTaskInputGenerator:
     def shared_to_aggregate(model_key):
         return [
             InputRef(
-                identifier=InputIdentifiers.models,
+                identifier=InputIdentifiers.shared,
                 parent_task_key=model_key,
                 parent_task_output_identifier=OutputIdentifiers.shared,
             )
@@ -310,11 +307,11 @@ class FLTaskOutputGenerator:
 
     @staticmethod
     def traintask(authorized_ids=None):
-        return {OutputIdentifiers.model: ComputeTaskOutputSpec(permissions=_permission_from_ids(authorized_ids))}
+        return {OutputIdentifiers.shared: ComputeTaskOutputSpec(permissions=_permission_from_ids(authorized_ids))}
 
     @staticmethod
     def aggregatetask(authorized_ids=None):
-        return {OutputIdentifiers.model: ComputeTaskOutputSpec(permissions=_permission_from_ids(authorized_ids))}
+        return {OutputIdentifiers.shared: ComputeTaskOutputSpec(permissions=_permission_from_ids(authorized_ids))}
 
     @staticmethod
     def predicttask(authorized_ids=None):